Refactor HW 3 and fix specifications.
