+++ .gitignore
@@ -17,3 +17,4 @@ buildNumber.properties
 .classpath
 *.log
 .checkstyle
+.vscode
