+++ pom.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>edu.byu.cs329</groupId>
@@ -9,42 +8,27 @@
 	<version>1.0-SNAPSHOT</version>
 
 	<name>HW3 MC/DC PDT Type-proof</name>
-	<!-- FIXME change it to the project's website -->
-	<url>http://www.example.com</url>
+	<url>https://github.com/byu-cs329/hw3-mcdc-pbt-typeproof</url>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<junit.jupiter.version>5.6.2</junit.jupiter.version>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<junit.jupiter.version>5.8.2</junit.jupiter.version>
 		<maven.compiler.version>3.8.1</maven.compiler.version>
 		<maven.site.version>3.9.1</maven.site.version>
-		<checkstyle.version>3.1.1</checkstyle.version>
+		<checkstyle.version>3.1.2</checkstyle.version>
 		<surefire.version>3.0.0-M5</surefire.version>
 		<javadoc.version>3.2.0</javadoc.version>
 		<maven.reports.version>3.1.0</maven.reports.version>
-		<jdt.core.version>3.7.1</jdt.core.version>
-		<log4j.slf4j.version>2.11.1</log4j.slf4j.version>
+		<log4j.slf4j.version>2.17.1</log4j.slf4j.version>
 		<exec.maven.version>3.0.0</exec.maven.version>
-		<junit.standalone.version>1.7.0-M1</junit.standalone.version>
+		<junit.standalone.version>1.8.2</junit.standalone.version>
 		<mockito.version>3.5.10</mockito.version>
-		<jacoco.version>0.8.6</jacoco.version>
+		<jacoco.version>0.8.7</jacoco.version>
 	</properties>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter</artifactId>
-			<version>${junit.jupiter.version}</version>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.eclipse.jdt</groupId>
-			<artifactId>org.eclipse.jdt.core</artifactId>
-			<version>${jdt.core.version}</version>
-		</dependency>
-
 		<dependency>
 			<groupId>org.apache.logging.log4j</groupId>
 			<artifactId>log4j-slf4j-impl</artifactId>
@@ -58,9 +42,17 @@
 		</dependency>
 
 		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter</artifactId>
+			<version>${junit.jupiter.version}</version>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
 			<groupId>org.mockito</groupId>
 			<artifactId>mockito-core</artifactId>
 			<version>${mockito.version}</version>
+			<scope>test</scope>
 		</dependency>
 
 		<dependency>
@@ -69,7 +61,6 @@
 			<version>${mockito.version}</version>
 			<scope>test</scope>
 		</dependency>
-		
 	</dependencies>
 
 	<build>
@@ -101,6 +92,13 @@
 					<failsOnError>true</failsOnError>
 					<linkXRef>false</linkXRef>
 				</configuration>
+				<dependencies>
+					<dependency>
+						<groupId>com.puppycrawl.tools</groupId>
+						<artifactId>checkstyle</artifactId>
+						<version>8.45.1</version>
+					</dependency>
+				</dependencies>
 				<executions>
 					<execution>
 						<id>validate</id>
@@ -187,7 +185,7 @@
 				<artifactId>maven-javadoc-plugin</artifactId>
 				<version>${javadoc.version}</version>
 				<configuration>
-					<source>8</source>
+					<source>11</source>
 					<show>public</show>
 					<tags>
 						<tag>
