--- src/main/resources/log4j.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration debug="true"
-  xmlns:log4j='http://jakarta.apache.org/log4j/'>
-
-	<appender name="console" class="org.apache.log4j.ConsoleAppender">
-	    <layout class="org.apache.log4j.PatternLayout">
-		<param name="ConversionPattern" 
-			value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n" />
-	    </layout>
-	</appender>
-
-	<appender name="file" class="org.apache.log4j.RollingFileAppender">
-	    <param name="append" value="false" />
-	    <param name="maxFileSize" value="10MB" />
-	    <param name="maxBackupIndex" value="10" />
-	    <param name="file" value="ConstantFold.log" />
-	    <layout class="org.apache.log4j.PatternLayout">
-		<param name="ConversionPattern" 
-			value="%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n" />
-	    </layout>
-	</appender>
-
-	<root>
-		<level value="INFO" />
-		<appender-ref ref="console" />
-		<appender-ref ref="file" />
-	</root>
-
-</log4j:configuration>
-
