+++ src/test/java/edu/byu/cs329/pbt/stateful/LlHashMapTest.java
@@ -1,7 +1,7 @@
 package edu.byu.cs329.pbt.stateful;
 
 /**
- * Test suite for the hash table.
+ * Test suite for the hash map.
  */
 public class LlHashMapTest {
